+++ phon/application/sound.hpp
@@ -93,7 +93,7 @@ public:
 
 	std::vector<double> get_intensity(intptr_t start_pos, intptr_t end_pos, double time_step);
 
-	Matrix<double> get_formants(double time, int nformant, double nyquist_frequency, double window_size, int lpc_order);
+	Array<double> get_formants(double time, int nformant, double nyquist_frequency, double window_size, int lpc_order);
 
 	static void initialize(Runtime &rt);