+++ docs/about/release-notes.rst
@@ -1,12 +1,13 @@
 Release notes
 -------------
 
-0.6.2 (??/11/2019)
+0.6.2 (29/10/2019)
 ~~~~~~~~~~~~~~~~~~
 
 - new concatenation operator ``&``
 - more ``Array`` methods
-- the multiplication and division operators can now operate on two arrays, or on an array and a a scalar value.
+- the multiplication and division operators can now operate on two arrays, or on an array and a scalar value.
+- comment/uncomment selection in script views
 
 0.6.1 (27/10/2019)
 ~~~~~~~~~~~~~~~~~~